Fix duplicate `KEEP_ALIVE` packets being sent out
Fix duplicate `KEEP_ALIVE` packets being sent out, because the check of `self.client_last_packet_received_time` alone was not enough. If the packet was sent out, as long as no response was received, the value remained the same and therefore again a `KEEP_ALIVE` packet was sent out.